# internal/interfaces/http/docs/v1_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "宠物养成游戏后端 REST API 文档",
        "title": "宠物养成游戏 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/pet": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "获取当前用户的宠物详细信息，包括外观、性格、技能和状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "获取我的宠物",
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pet.PetDetailDTO"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "还没有宠物",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "创建一个新宠物，可以指定物种或随机生成",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "创建宠物",
                "parameters": [
                    {
                        "description": "创建宠物请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pet.CreatePetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pet.CreatePetResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "409": {
                        "description": "已拥有宠物",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/pet/clean": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "清洁宠物，增加清洁度",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "清洁宠物",
                "responses": {
                    "200": {
                        "description": "清洁成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pet.CleanPetResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/pet/feed": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "使用食物道具喂食宠物，增加饱食度和经验值，可能升级",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "喂食宠物",
                "parameters": [
                    {
                        "description": "喂食请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pet.FeedPetRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "喂食成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pet.FeedPetResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
        "/pet/play": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "与宠物互动玩耍，增加快乐度和经验值，消耗精力",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pet"
                ],
                "summary": "和宠物玩耍",
                "responses": {
                    "200": {
                        "description": "玩耍成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/pet.PlayPetResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "500": {
                        "description": "服务器错误",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pet.AppearanceDTO": {
            "type": "object",
            "properties": {
                "bodyType": {
                    "type": "string"
                },
                "colorPrimary": {
                    "type": "string"
                },
                "colorSecondary": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "patternType": {
                    "type": "string"
                }
            }
        },
        "pet.CleanPetResponse": {
            "type": "object",
            "properties": {
                "cleanliness": {
                    "description": "当前清洁度",
                    "type": "integer"
                }
            }
        },
        "pet.CreatePetRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "description": "宠物名称",
                    "type": "string",
                    "maxLength": 20,
                    "minLength": 1
                },
                "speciesId": {
                    "description": "物种ID（可选，不填则随机）",
                    "type": "string"
                }
            }
        },
        "pet.CreatePetResponse": {
            "type": "object",
            "properties": {
                "pet": {
                    "$ref": "#/definitions/pet.PetDetailDTO"
                }
            }
        },
        "pet.FeedPetRequest": {
            "type": "object",
            "required": [
                "foodItemId"
            ],
            "properties": {
                "foodItemId": {
                    "description": "食物道具ID",
                    "type": "integer"
                }
            }
        },
        "pet.FeedPetResponse": {
            "type": "object",
            "properties": {
                "expGained": {
                    "description": "获得经验",
                    "type": "integer"
                },
                "hunger": {
                    "description": "当前饱食度",
                    "type": "integer"
                },
                "levelUp": {
                    "description": "是否升级",
                    "type": "boolean"
                },
                "newLevel": {
                    "description": "新等级（如果升级）",
                    "type": "integer"
                }
            }
        },
        "pet.PersonalityDTO": {
            "type": "object",
            "properties": {
                "activity": {
                    "type": "integer"
                },
                "appetite": {
                    "type": "integer"
                },
                "curiosity": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "social": {
                    "type": "integer"
                }
            }
        },
        "pet.PetDetailDTO": {
            "type": "object",
            "properties": {
                "appearance": {
                    "description": "外观",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pet.AppearanceDTO"
                        }
                    ]
                },
                "exp": {
                    "description": "当前经验",
                    "type": "integer"
                },
                "expToNext": {
                    "description": "升级所需经验",
                    "type": "integer"
                },
                "geneCode": {
                    "description": "基因码（可选，用于展示独特性）",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "level": {
                    "description": "等级",
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "personality": {
                    "description": "性格",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pet.PersonalityDTO"
                        }
                    ]
                },
                "skill": {
                    "description": "技能",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pet.SkillDTO"
                        }
                    ]
                },
                "stage": {
                    "description": "成长",
                    "type": "string"
                },
                "status": {
                    "description": "状态",
                    "allOf": [
                        {
                            "$ref": "#/definitions/pet.StatusDTO"
                        }
                    ]
                }
            }
        },
        "pet.PlayPetResponse": {
            "type": "object",
            "properties": {
                "energy": {
                    "description": "当前精力",
                    "type": "integer"
                },
                "expGained": {
                    "description": "获得经验",
                    "type": "integer"
                },
                "happiness": {
                    "description": "当前快乐度",
                    "type": "integer"
                },
                "levelUp": {
                    "description": "是否升级",
                    "type": "boolean"
                }
            }
        },
        "pet.SkillDTO": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "rarity": {
                    "type": "string"
                }
            }
        },
        "pet.StatusDTO": {
            "type": "object",
            "properties": {
                "cleanliness": {
                    "type": "integer"
                },
                "energy": {
                    "type": "integer"
                },
                "happiness": {
                    "type": "integer"
                },
                "hunger": {
                    "type": "integer"
                },
                "isDirty": {
                    "type": "boolean"
                },
                "isHungry": {
                    "type": "boolean"
                },
                "isTired": {
                    "type": "boolean"
                },
                "isUnhappy": {
                    "type": "boolean"
                }
            }
        },
...
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}